refactor: 리뷰 페이지
* 타입 파일 이동 및 수정
* 유틸 함수 파일명 수정
* Card 테스트 수정
