+++ package.json
@@ -20,7 +20,6 @@
     "@hookform/resolvers": "^3.9.0",
     "@tanstack/react-query": "^5.51.24",
     "@tanstack/react-query-devtools": "^5.51.24",
-    "@types/lodash": "^4.17.7",
     "axios": "^1.7.4",
     "jotai": "^2.9.3",
     "lodash": "^4.17.21",
@@ -48,6 +47,7 @@
     "@testing-library/jest-dom": "^6.5.0",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
+    "@types/lodash": "^4.17.7",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^4.3.1",
