+++ src/pages/Reviews/components/Card.test.tsx
@@ -11,44 +11,16 @@ describe('Card Components', () => {
    * Card 컴포넌트가 렌더 되는지 확인하는 테스트를 작성합니다.
    */
   describe('Card 컴포넌트가 렌더되어야한다.', () => {
-    it('Card contents props가 렌더링 되어야한다.', () => {
+    it('Card 모든 props가 렌더링 되어야한다.', () => {
       render(
         <Card {...cardData} />
       );
 
+      expect(screen.getByText(cardData.reviewDate)).toBeInTheDocument();
       expect(screen.getByText(cardData.content)).toBeInTheDocument();
-    })
-
-    it('Card reviewDate props가 렌더링 되어야한다.', () => {
-      render(
-        <Card {...cardData} />
-      );
-
-      expect(screen.getByText(cardData.reviewDate)).toBeInTheDocument();
-    })
-
-    it('Card reviewDate props가 렌더링 되어야한다.', () => {
-      render(
-        <Card {...cardData} />
-      );
-
-      expect(screen.getByText(cardData.reviewDate)).toBeInTheDocument();
-    })
-
-    it('Card menu props가 렌더링 되어야한다.', () => {
-      render(
-        <Card {...cardData} />
-      );
-
-      expect(screen.getByText(cardData.menus.name)).toBeInTheDocument();
-    })
-
-    it('Card members props가 렌더링 되어야한다.', () => {
-      render(
-        <Card {...cardData} />
-      );
-
       expect(screen.getByText(cardData.member.nickname)).toBeInTheDocument();
+      expect(screen.getByText(cardData.menus.name)).toBeInTheDocument();
+      expect(screen.getByText(cardData.content)).toBeInTheDocument();
     })
   });
 })
